bootstrap {renv}	R Documentation

Bootstrap an renv Installation

Description

Bootstrap an renv installation, making the requested version of renv available for projects on the system.

Usage

bootstrap(project = NULL, version = NULL)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

version
The version of renv to install. If NULL, the version of renv currently installed will be used. The requested version of renv will be retrieved from the renv public GitHub repository, at https://github.com/rstudio/renv.

Details

Normally, this function does not need to be called directly by the user; it will be invoked as required by init() and activate().

Value

The project directory, invisibly. Note that this function is normally called for its side effects.

[Package renv version 0.9.2 Index]