checkout {renv}	R Documentation

Checkout a Repository

Description

renv::checkout() can be used to install and use the latest packages available from the requested repositories. This can be useful for cleaning up a library which has become a mish-mash of packages installed from a variety of disparate sources.

Usage


checkout(
  repos = getOption("repos"),
  ...,
  packages = NULL,
  clean = FALSE,
  project = NULL
)

Arguments

repos
The R package repositories to check out.

...
Unused arguments, reserved for future expansion. If any arguments are matched to ..., renv will signal an error.

packages
The packages to be installed. When NULL (the default), all packages currently used in the project will be installed.

clean
Boolean; remove packages not recorded in the lockfile from the target library? Use clean = TRUE if you'd like the library state to exactly reflect the lockfile contents after restore().

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

[Package renv version 0.13.2-16 Index]