deactivate {renv}	R Documentation

Deactivate a Project

Description

Use deactivate() to remove the infrastructure used by renv to activate projects for newly-launched R sessions. In particular, this implies removing the requisite code from the project .Rprofile that automatically activates the project when new R sessions are launched in the project directory.

Usage

deactivate(project = NULL)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Value

The project directory, invisibly. Note that this function is normally called for its side effects.

See Also

Other renv: activate()

Examples

## Not run: 

# deactivate the currently-activated project
renv::deactivate()

## End(Not run)

[Package renv version 0.9.2 Index]