equip {renv}	R Documentation

Install required system libraries

Description

Equip your system with libraries commonly-used during compilation of base and recommended R packages. This was previously useful with older versions of R on windows, but is no longer terribly helpful.

Usage

equip()

Value

This function is normally called for its side effects.

Examples

## Not run: 

# download useful build tools
renv::equip()

## End(Not run)

[Package renv version 1.1.7 Index]