renv-package {renv}	R Documentation

Project-local Environments for R

Description

Project-local environments for R.

Details

You can use renv to construct isolated, project-local R libraries. Each project using renv will share package installations from a global cache of packages, helping to avoid wasting disk space on multiple installations of a package that might otherwise be shared across projects.

Author(s)

Maintainer: Kevin Ushey kevin@rstudio.com

Other contributors:

Posit Software, PBC [copyright holder, funder]

See Also

Useful links:

https://rstudio.github.io/renv/

Report bugs at https://github.com/rstudio/renv/issues

[Package renv version 0.17.0 Index]