Generate renv.lock from an RStudio Connect manifest.json

Use renv_lockfile_from_manifest() to convert a manifest.json file from an RStudio Connect content bundle into an renv.lock lockfile.

renv_lockfile_from_manifest(manifest, lockfile = NA)

Arguments

manifest
The path to a manifest.json file.

lockfile
The path to the lockfile to be generated and / or updated. When NA (the default), the generated lockfile is returned as an R object; otherwise, the lockfile will be written to the path specified by lockfile.

Value

An renv lockfile.

Details

This function can be useful when you need to recreate the package environment of a piece of content that is deployed to RStudio Connect. The content bundle contains a manifest.json file that is used to recreate the package environment. This function will let you convert that manifest file to an renv.lock file. Run renv::restore() after you've converted the file to restore the package environment.

By default the lockfile argument is set to NA. This will not create a new renv.lock file. Rather, it will return a lockfile object (see ?lockfile) that can be used to create a new renv.lock file. If lockfile is set to a character string, a new file will be created with that path -- e.g. renv.lock -- and the lockfile object will be returned.