renv_use_python_virtualenv_impl {renv}	R Documentation

Internal helper for activating a Python virtual environment

Description

Internal helper for activating a Python virtual environment

Usage

renv_use_python_virtualenv_impl(
  project,
  name = NULL,
  version = NULL,
  python = NULL
)

Arguments

project
The project directory.

name
The environment name, if any. If unset, it should be constructed based on the Python executable used (note: not the version parameter)

version
The requested version of Python (which may not be the actual version!) This version should be used as a hint for finding an appropriate version of Python, if the environment needs to be re-created.

python
The copy of Python to be used. When unset, an appropriate version of Python should be discovered based on the version parameter.

Value

The path to the Python binary in the associated virtual environment.

[Package renv version 0.14.0-79 Index]