activate {renv}	R Documentation

Activate a Project

Description

Use activate() to write the infrastructure needed to ensure that newly-launched R projects will load the project's private library on launch, alongside any other project-specific state recorded for the project.

Usage

activate(project = NULL)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Value

The project directory, invisibly. Note that this function is normally called for its side effects.

See Also

Other renv: deactivate()

Examples

## Not run: 

# activate the current project
renv::activate()

# activate a separate project
renv::activate("~/projects/analysis")

## End(Not run)

[Package renv version 0.9.2 Index]