activate {renv}	R Documentation

Activate a Project

Description

Activate a project, thereby loading it in the current session and also writing the infrastructure necessary to ensure the project is auto-loaded for newly-launched R sessions.

Usage

activate(project = NULL, profile = NULL)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

profile
The profile to be activated. When NULL, the default profile is activated instead. See vignette("profiles", package = "renv") for more information.

Details

Using activate() will:

Load the requested project via load(),

Add source("renv/activate.R") to the project .Rprofile, thereby instructing newly-launched R sessions to automatically load the current project.

Normally, activate() is called as part of init() when a project is first initialized. However, activate() can be used to activate (or re-activate) an renv project – for example, if the project was shared without the auto-loader included in the project .Rprofile, or because that project was previously deactivated (via deactivate()).

Value

The project directory, invisibly. Note that this function is normally called for its side effects.

See Also

Other renv: deactivate()

Examples

## Not run: 

# activate the current project
renv::activate()

# activate a separate project
renv::activate("~/projects/analysis")

## End(Not run)

[Package renv version 0.17.0 Index]