clean {renv}	R Documentation

Clean a Project

Description

Clean up a project and its associated R libraries.

Usage

clean(project = NULL, ..., confirm = interactive())

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

...
Optional arguments; reserved for future expansion.

confirm
Boolean; prompt the user before taking any action?

Details

The following actions will be taken:

Stale lockfiles (00LOCK-) will be removed.

Leftover temporary directories in the project library will be removed.

Non-system packages installed in the system library will be removed.

Unused packages within the project will be removed.

Packages within the cache that are no longer used will be removed.

Value

The project directory, invisibly. Note that this function is normally called for its side effects.

Examples

## Not run: 

# clean the current project
renv::clean()

## End(Not run)

[Package renv version 0.9.2 Index]