User-Level Configuration of renv

Configure different behaviors of renv.

Details

For a given configuration option:

If an R option of the form renv.config.<name> is available, then that option's value will be used;

If an environment variable of the form RENV_CONFIG_<NAME> is available, then that option's value will be used;

Otherwise, the default for that particular configuration value is used.

Any periods (.)s in the option name are transformed into underscores (_) in the environment variable name, and vice versa. For example, the configuration option auto.snapshot could be configured as:

options(renv.config.auto.snapshot = <...>)

Sys.setenv(RENV_CONFIG_AUTO_SNAPSHOT = <...>)

Note that if both the R option and the environment variable are defined, the R option will be used instead. Environment variables can be more useful when you want a particular configuration to be automatically inherited by child processes; if that behavior is not desired, then the R option may be preferred.

If you want to set and persist these options across multiple projects, it is recommended that you set them in your user startup files (e.g. in ~/.Rprofile).

Configuration

The following renv configuration options are available:

Name	Description	auto.snapshot

logical(1)	Automatically snapshot changes to the project library after a new package is installed? Note that package upgrades or removals will not be automatically snapshotted. (Boolean; defaults to TRUE)	use.cache

logical(1)	Use the global cache when installing packages? (Boolean; defaults to TRUE)