consent {renv}	R Documentation

Consent to usage of renv

Description

Provide consent to renv, allowing it to write and update certain files on your filesystem.

Usage

consent()

Details

As part of its normal operation, renv will write and update some files in your project directory, as well as an application-specific cache directory. These paths are documented within paths.

In accordance with the CRAN Repository Policy, renv must first obtain consent from you, the user, before these actions can be taken. Please call renv::consent() in an interactive session to provide this consent.

You can also set the R option:

options(renv.consent = TRUE)

to implicitly provide this consent for e.g. non-interactive R sessions.

[Package renv version 0.7.0-15 Index]