consent {renv}	R Documentation

Consent to usage of renv

Description

Provide consent to renv, allowing it to write and update certain files on your filesystem.

Usage

consent(provided = FALSE)

Arguments

provided
The default provided response. If you need to provide consent from a non-interactive R session, you can invoke renv::consent(provided = TRUE) explicitly.

Details

As part of its normal operation, renv will write and update some files in your project directory, as well as an application-specific cache directory. These paths are documented within paths.

In accordance with the CRAN Repository Policy, renv must first obtain consent from you, the user, before these actions can be taken. Please call renv::consent() first to provide this consent.

You can also set the R option:

options(renv.consent = TRUE)

to implicitly provide consent for e.g. non-interactive R sessions.

Value

TRUE if consent is provided, or an R error otherwise.

[Package renv version 0.9.2 Index]