history {renv}	R Documentation

View Lockfile History

Description

Use your version control system to find prior versions of the renv.lock file that have been used in your project.

Usage

history(project = NULL)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Details

The history() function is currently only implemented for projects using git for version control.

Value

An R data.frame, summarizing the commits in which renv.lock has been mutated.

Examples

## Not run: 

# get history of previous versions of renv.lock in VCS
db <- renv::history()

# choose an older commit
commit <- db$commit[5]

# revert to that version of the lockfile
renv::revert(commit = commit)

## End(Not run)

[Package renv version 0.9.2 Index]