hydrate {renv}	R Documentation

Hydrate a Project

Description

Discover the R packages used within a project, and then install those packages into the active library. This effectively allows you to clone the state of your system R libraries for use within a project library.

Usage

hydrate(packages = NULL, ..., library = NULL, project = NULL)

Arguments

packages
The set of R packages to install. When NULL, the set of packages as reported by dependencies() is used.

...
Optional arguments; reserved for future expansion.

library
The R library to be hydrated. When NULL, the active library as reported by .libPaths() is used.

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Details

hydrate() attempts to re-use packages already installed on your system, to avoid unnecessary attempts to download and install packages from remote sources. hydrate() will attempt to discover R packages from the following sources (in order):

The user library,

The site library,

The system library,

The renv cache.

If package is discovered in one of these locations, renv will attempt to copy or link that package into the requested library as appropriate.

Value

A named R list, giving the packages that were used for hydration as well as the set of packages which were not found.

Missing Packages

If renv discovers that your project depends on R packages not currently installed in your user library, then it will attempt to install those packages from the active R repositories.

Examples

## Not run: 

# hydrate the active library
renv::hydrate()

## End(Not run)

[Package renv version 0.9.2 Index]