install {renv}	R Documentation

Install Packages

Description

Install one or more R packages from a variety of remote sources.

Usage

install(
  packages = NULL,
  ...,
  library = NULL,
  rebuild = FALSE,
  confirm = interactive(),
  project = NULL
)

Arguments

packages
A character vector of R packages to install. Required package dependencies (Depends, Imports, LinkingTo) will be installed as required.

...
Optional arguments; reserved for future expansion.

library
The R library to be used. When NULL, the active project library will be used instead.

rebuild
Force packages to be rebuilt, thereby bypassing any installed versions of the package available in the cache? This can either be a boolean (indicating that the requested package(s) should be rebuilt), or a vector of package names indicating which packages should be rebuilt.

confirm
Boolean; prompt the user before taking any action?

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Details

install() uses the same machinery as restore() when installing packages. In particular, this means that the local cache of package installations is used when possible. This helps to avoid re-downloading packages that have already been downloaded before, and re-compiling packages from source when a binary copy of that package is already available.

Note that this interface is subject to change – the goal is to hook into separate package installation backends in the future.

Value

A named list of package records which were installed by renv.

Examples

## Not run: 

# install the latest version of 'digest'
renv::install("digest")

# install an old version of 'digest' (using archives)
renv::install("digest@0.6.18")

# install 'digest' from GitHub (latest dev. version)
renv::install("eddelbuettel/digest")

# install a package from local sources
renv::install("~/path/to/package")

## End(Not run)

[Package renv version 0.9.2 Index]