modify {renv}	R Documentation

Open the Lockfile for Editing

Description

Open a project's lockfile (if any) for editing. After edit, if the lockfile edited is associated with the active project, any state-related changes (e.g. to R repositories) will be updated in the current session.

Usage

modify(project = NULL)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Value

The project directory, invisibly. Note that this function is normally called for its side effects.

Examples

## Not run: 

# modify an existing lockfile
if (interactive())
  renv::modify()

## End(Not run)

[Package renv version 0.9.2 Index]