rebuild {renv}	R Documentation

Rebuild the Packages in your Project Library

Description

Rebuild and reinstall packages in your library. This can be useful as a diagnostic tool – for example, if you find that one or more of your packages fail to load, and you want to ensure that you are starting from a clean slate.

Usage

rebuild(
  packages = NULL,
  recursive = TRUE,
  ...,
  confirm = interactive(),
  library = NULL,
  project = NULL
)

Arguments

packages
The package(s) to be rebuilt. When NULL, all packages in the library will be installed.

recursive
Boolean; should dependencies of packages be rebuilt recursively? Defaults to TRUE.

...
Optional arguments; reserved for future expansion.

confirm
Boolean; prompt the user before taking any action?

library
The R library to be used. When NULL, the active project library will be used instead.

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Details

Note that binaries will be used when appropriate and available for your platform. If you'd like to force packages to be rebuilt from sources, you can set options(pkgType = "source").

Value

A named list of package records which were installed by renv.

Examples

## Not run: 

# rebuild the 'dplyr' package + all of its dependencies
renv::rebuild("dplyr", recursive = TRUE)

# rebuild only 'dplyr'
renv::rebuild("dplyr", recursive = FALSE)

## End(Not run)

[Package renv version 0.9.2 Index]