rehash {renv}	R Documentation

Re-Hash Packages in the renv Cache

Description

Re-hash packages in the renv cache, ensuring that any previously-cached packages are copied to a new cache location appropriate for this version of renv. This can be useful if the cache scheme has changed in a new version of renv, but you'd like to preserve your previously-cached packages.

Usage

rehash(confirm = interactive())

Arguments

confirm
Boolean; prompt the user before taking any action?

Details

Any packages which are re-hashed will retain links to the location of the newly-hashed package, ensuring that prior installations of renv can still function as expected.

[Package renv version 0.9.2 Index]