remote {renv}	R Documentation

Resolve a Remote

Description

Given a remote specification, resolve it into an renv package record that can be used for download and installation (e.g. with install).

Usage

remote(spec)

Arguments

spec
A remote specification.

[Package renv version 0.9.2 Index]