remove {renv}	R Documentation

Remove Packages

Description

Remove (uninstall) R packages.

Usage

remove(packages, ..., library = NULL, project = NULL)

Arguments

packages
A character vector of R packages to remove.

...
Unused arguments, reserved for future expansion. If any arguments are matched to ..., renv will signal an error.

library
The library from which packages should be removed. When NULL, the active library (that is, the first entry reported in .libPaths()) is used instead.

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Value

A vector of package records, describing the packages (if any) which were successfully removed.

Examples

## Not run: 
##D 
##D # disable automatic snapshots
##D auto.snapshot <- getOption("renv.config.auto.snapshot")
##D options(renv.config.auto.snapshot = FALSE)
##D 
##D # initialize a new project (with an empty R library)
##D renv::init(bare = TRUE)
##D 
##D # install digest 0.6.19
##D renv::install("digest@0.6.19")
##D 
##D # save library state to lockfile
##D renv::snapshot()
##D 
##D # remove digest from library
##D renv::remove("digest")
##D 
##D # check library status
##D renv::status()
##D 
##D # restore lockfile, thereby reinstalling digest 0.6.19
##D renv::restore()
##D 
##D # restore automatic snapshots
##D options(renv.config.auto.snapshot = auto.snapshot)
##D 
## End(Not run)

[Package renv version 0.14.0-79 Index]