Project-local Environments for R

Project-local environments for R.

Details

You can use renv to construct isolated, project-local R libraries. Each project using renv will share package installations from a global cache of packages, helping to avoid wasting disk space on multiple installations of a package that might otherwise be shared across projects.

See also

Author

Maintainer: Kevin Ushey kevin@rstudio.com (ORCID)

Authors:

Hadley Wickham hadley@rstudio.com (ORCID)

Other contributors:

Posit Software, PBC [copyright holder, funder]