repair {renv}	R Documentation

Repair a Project Library

Description

Repair a project library whose cache symlinks have become broken. renv will attempt to reinstall the requisite packages.

Usage

repair(library = NULL, lockfile = NULL, project = NULL)

Arguments

library
The R library to be used. When NULL, the active project library will be used instead.

lockfile
The path to a lockfile (if any). When available, renv will use the lockfile when attempting to infer the remote associated with the inaccessible version of each missing package.

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

[Package renv version 0.17.0 Index]