scaffold {renv}	R Documentation

Generate renv Project Infrastructure

Description

Write the renv project infrastructure for a project.

Usage


scaffold(project = NULL, version = NULL)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

version
The version of renv to associate with this project. By default, the version of renv currently installed is used.

Details

The function will update the project .Rprofile so that renv is automatically loaded for new R sessions launched in this project. renv will also be installed and made available within the project's private library.

[Package renv version 0.9.2-18 Index]