scaffold {renv}	R Documentation

Generate renv Project Infrastructure

Description

Write the renv project infrastructure for a project.

Usage

scaffold(
  project = NULL,
  version = NULL,
  repos = getOption("repos"),
  settings = NULL
)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

version
The version of renv to associate with this project. By default, the version of renv currently installed is used.

repos
The R repositories to associate with this project.

settings
A list of renv settings, to be applied to the project after creation. These should map setting names to the desired values. See settings for more details.

Details

Invoking renv::scaffold() will:

Install renv into the project library,

Update the project .Rprofile so that renv is automatically loaded for new R sessions launched in this project, and

Write a bare lockfile renv.lock.

Examples

## Not run: 
# create scaffolding with 'devtools' ignored
renv::scaffold(settings = list(ignored.packages = "devtools"))

## End(Not run)

[Package renv version 0.17.0 Index]