Project Settings

Define project-local settings that can be used to adjust the behavior of renv with your particular project.

settings

Format

An object of class list of length 4.

ignored.packages

c("", "A vector of packages, which should be ignored when attempting to snapshot\nthe project's private library. Note that if a package has already been\nadded to the lockfile, that entry in the lockfile will not be ignored.
")

external.libraries

c("", "A vector of library paths, to be used in addition to the project's own\nprivate library. This can be useful if you have a package available for use\nin some global library, but for some reason renv is not able to install\nthat package (e.g. sources or binaries for that package are not publicly\navailable, or you have been unable to orchestrate the pre-requisites for\ninstalling some packages from source on your machine).")

snapshot.type

c("", "The type of snapshot to perform by default. See snapshot for more\ndetails.")

use.cache

c("", "Use a global cache of R packages. When active, renv will install\npackages into a global cache, and link packages from the cache into your\nrenv projects as appropriate. This can greatly save on disk space\nand install time when for R packages which are used across multiple\nprojects in the same environment.")

Defaults

You can change the default values of these settings for newly-created renv projects by setting R options for renv.settings or renv.settings.<name>. For example:

options(renv.settings = list(snapshot.type = "simple"))
options(renv.settings.snapshot.type = "simple")

If both of the renv.settings and renv.settings.<name> options are set for a particular key, the option associated with renv.settings.<name> is used instead. We recommend setting these in an appropriate startup profile, e.g. ~/.Rprofile or similar.

Examples

# NOT RUN {
# view currently-ignored packaged
renv::settings$ignored.packages()

# ignore a set of packages
renv::settings$ignored.packages("devtools", persist = FALSE)

# }