status {renv}	R Documentation

Status

Description

Report differences between the project's lockfile and the current state of the project's library (if any).

Usage

status(project = NULL, ..., library = NULL, lockfile = NULL, cache = FALSE)

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

...
Unused arguments, reserved for future expansion. If any arguments are matched to ..., renv will signal an error.

library
The library paths. By default, the library paths associated with the requested project are used.

lockfile
The path to a lockfile. By default, the project lockfile (called renv.lock) is used.

cache
Boolean; perform diagnostics on the global package cache? When TRUE, renv will validate that the packages installed into the cache are installed at the expected + proper locations, and validate the hashes used for those storage locations.

Value

This function is normally called for its side effects.

Examples

## Not run: 
##D 
##D # disable automatic snapshots
##D auto.snapshot <- getOption("renv.config.auto.snapshot")
##D options(renv.config.auto.snapshot = FALSE)
##D 
##D # initialize a new project (with an empty R library)
##D renv::init(bare = TRUE)
##D 
##D # install digest 0.6.19
##D renv::install("digest@0.6.19")
##D 
##D # save library state to lockfile
##D renv::snapshot()
##D 
##D # remove digest from library
##D renv::remove("digest")
##D 
##D # check library status
##D renv::status()
##D 
##D # restore lockfile, thereby reinstalling digest 0.6.19
##D renv::restore()
##D 
##D # restore automatic snapshots
##D options(renv.config.auto.snapshot = auto.snapshot)
##D 
## End(Not run)

[Package renv version 0.14.0-79 Index]