update {renv}	R Documentation

Update Packages

Description

Update packages which are currently out-of-date. Currently, only CRAN and GitHub package sources are supported.

Usage

update(
  packages = NULL,
  ...,
  library = NULL,
  rebuild = FALSE,
  check = FALSE,
  confirm = interactive(),
  project = NULL
)

Arguments

packages
A character vector of R packages to update. When NULL, all packages within the required libraries will be updated.

...
Optional arguments; reserved for future expansion.

library
The R library to be used. When NULL, the active project library will be used instead.

rebuild
Force packages to be rebuilt, thereby bypassing any installed versions of the package available in the cache? This can either be a boolean (indicating that the requested package(s) should be rebuilt), or a vector of package names indicating which packages should be rebuilt.

check
Boolean; check for package updates without actually installing available updates?

confirm
Boolean; prompt the user before taking any action?

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

Details

Updates will only be checked from the same source – for example, if a package was installed from GitHub, but a newer version is available on CRAN, that updated version will not be seen.

Value

A named list of package records which were installed by renv.

Examples

## Not run: 

# update the 'dplyr' package
renv::update("dplyr")

## End(Not run)

[Package renv version 0.9.2 Index]