upgrade {renv}	R Documentation

Upgrade renv

Description

Upgrade the version of renv associated with a project.

Usage

upgrade(project = NULL, version = NULL, reload = NULL, confirm = interactive())

Arguments

project
The project directory. If NULL, then the active project will be used. If no project is currently active, then the current working directory is used instead.

version
The version of renv to be installed. By default, the latest version of renv as available on the active R package repositories is used.

reload
Boolean; reload renv after install? When NULL (the default), renv will be re-loaded only if updating renv for the active project. Note that this may fail if you've loaded packages which also depend on renv.

confirm
Boolean; confirm upgrade before proceeding?

Details

By default, this function will attempt to install the latest version of renv as available on the active R package repositories. If you'd instead like to try out a development version of renv, you can explicitly request a different version of renv and that version of the package will be downloaded and installed from GitHub. Use version = "master" to install the latest development version of renv, as from the renv project's GitHub page.

Value

A boolean value, indicating whether the requested version of renv was successfully installed. Note that this function is normally called for its side effects.

Examples

## Not run: 

# upgrade to the latest version of renv
renv::upgrade()

# upgrade to the latest version of renv on GitHub (development version)
renv::upgrade(version = "master")

## End(Not run)

[Package renv version 0.9.2 Index]